+++ packages/react-cosmos-playground2/src/plugins/RendererCore/__tests__/isRendererConnected.ts
@@ -1,7 +1,8 @@
 import { NotificationsSpec } from '../../Notifications/public';
 import { loadPlugins } from 'react-plugin';
 import { cleanup, mockMethodsOf } from '../../../testHelpers/plugin';
-import { getRendererCoreMethods, mockRendererReady } from '../testHelpers';
+import { getRendererCoreMethods } from '../../../testHelpers/pluginMocks';
+import { mockRendererReady } from '../testHelpers';
 import { register } from '..';
 import { RouterSpec } from '../../Router/public';
