+++ packages/react-cosmos-playground2/src/plugins/RendererCore/__tests__/methods.tsx
@@ -1,11 +1,8 @@
 import { loadPlugins } from 'react-plugin';
 import { cleanup, mockMethodsOf } from '../../../testHelpers/plugin';
+import { getRendererCoreMethods } from '../../../testHelpers/pluginMocks';
 import { NotificationsSpec } from '../../Notifications/public';
-import {
-  mockRendererReady,
-  getRendererCoreMethods,
-  mockFixtureStateChange
-} from '../testHelpers';
+import { mockRendererReady, mockFixtureStateChange } from '../testHelpers';
 import { RouterSpec } from '../../Router/public';
 import { register } from '..';
