+++ packages/react-cosmos-playground2/src/plugins/RendererCore/__tests__/selectPrimaryRenderer.ts
@@ -3,7 +3,8 @@ import { loadPlugins } from 'react-plugin';
 import { NotificationsSpec } from '../../Notifications/public';
 import { RouterSpec } from '../../Router/public';
 import { cleanup, mockMethodsOf } from '../../../testHelpers/plugin';
-import { getRendererCoreMethods, mockRendererReady } from '../testHelpers';
+import { getRendererCoreMethods } from '../../../testHelpers/pluginMocks';
+import { mockRendererReady } from '../testHelpers';
 import { register } from '..';
 
 afterEach(cleanup);
