+++ packages/react-cosmos-playground2/src/plugins/RendererCore/receiveResponse/__tests__/fixtureStateChange.ts
@@ -8,11 +8,11 @@ import {
   mockMethodsOf,
   on
 } from '../../../../testHelpers/plugin';
+import { getRendererCoreMethods } from '../../../../testHelpers/pluginMocks';
 import { RouterSpec } from '../../../Router/public';
 import {
   createFixtureStateChangeResponse,
-  mockRendererReady,
-  getRendererCoreMethods
+  mockRendererReady
 } from '../../testHelpers';
 import { RendererCoreSpec } from '../../public';
 import { register } from '../..';
